feat: Update and Combine Rounter for Journal Section 📚 (#4089)
