+++ docs/en/journal.md
@@ -8,15 +8,14 @@ pageClass: routes
 
 ### Cell Journal
 
-<Route author="yech1990" example="/cell/cell/current" path="/cell/cell/:category" />
+<RouteEn author="yech1990" example="/cell/cell/current" path="/cell/cell/:category" />
 
-> Current Issue (default)
+| `:category` |       Query Type        | Route                                    |
+| :---------: | :---------------------: | ---------------------------------------- |
+|   current   | Current Issue (default) | [/cell/cell/current](/cell/cell/current) |
+|   inpress   |    Articles in press    | [/cell/cell/inpress](/cell/cell/inpress) |
 
-`/cell/cell/current`
-
-> Articles in press
-
-`/cell/cell/inpress`
+</RouteEn>
 
 ### eLife - Latest Research-ALL
 
@@ -28,7 +27,46 @@ pageClass: routes
 
 ### Nature Journal - Latest Research
 
-<RouteEn author="yech1990" example="/nature/nature/research" path="/nature/nature/research" />
+<RouteEn author="yech1990" example="/nature/research/ng" path="/nature/research/:journal" :paramsDesc="['short name for a journal']" />
+
+|  `:journal`   |  Full Name of the Journal   | Route                                                            |
+| :-----------: | :-------------------------: | ---------------------------------------------------------------- |
+|    nature     |           Nature            | [/nature/research/nature](/nature/research/nature)               |
+|      nbt      |    Nature Biotechnology     | [/nature/research/nbt](/nature/research/nbt)                     |
+|     neuro     |     Nature Neuroscience     | [/nature/research/neuro](/nature/research/neuro)                 |
+|      ng       |       Nature Genetics       | [/nature/research/ng](/nature/research/ng)                       |
+|      ni       |      Nature Immunology      | [/nature/research/ni](/nature/research/ni)                       |
+|     nmeth     |        Nature Method        | [/nature/research/nmeth](/nature/research/nmeth)                 |
+|     nchem     |      Nature Chemistry       | [/nature/research/nchem](/nature/research/nchem)                 |
+|     nmat      |      Nature Materials       | [/nature/research/nmat](/nature/research/nmat)                   |
+| natmachintell | Nature Machine Intelligence | [/nature/research/natmachintell](/nature/research/natmachintell) |
+
+-   Using router (`/nature/research/` + “short name for a journal”) to query latest research paper for a certain journal of Nature Publishing Group.
+    If the `:journal` parameter is blank, then latest research of Nature will return.
+-   The journals from NPG are run by different group of people, and the website of may not be consitent for all the journals
+-   Only the abstract section is rendered
+
+</RouteEn>
+
+### Nature Journal - News & Comment
+
+<RouteEn author="yech1990" example="/nature/news-and-comment/ng" path="/nature/news-and-comment/:journal" :paramsDesc="['short name for a journal']" />
+
+|  `:journal`   |  Full Name of the Journal   | Route                                                                            |
+| :-----------: | :-------------------------: | -------------------------------------------------------------------------------- |
+|      nbt      |    Nature Biotechnology     | [/nature/news-and-comment/nbt](/nature/news-and-comment/nbt)                     |
+|     neuro     |     Nature Neuroscience     | [/nature/news-and-comment/neuro](/nature/news-and-comment/neuro)                 |
+|      ng       |       Nature Genetics       | [/nature/news-and-comment/ng](/nature/news-and-comment/ng)                       |
+|      ni       |      Nature Immunology      | [/nature/news-and-comment/ni](/nature/news-and-comment/ni)                       |
+|     nmeth     |        Nature Method        | [/nature/news-and-comment/nmeth](/nature/news-and-comment/nmeth)                 |
+|     nchem     |      Nature Chemistry       | [/nature/news-and-comment/nchem](/nature/news-and-comment/nchem)                 |
+|     nmat      |      Nature Materials       | [/nature/news-and-comment/nmat](/nature/news-and-comment/nmat)                   |
+| natmachintell | Nature Machine Intelligence | [/nature/news-and-comment/natmachintell](/nature/news-and-comment/natmachintell) |
+
+-   Using router (`/nature/research/` + “short name for a journal”) to query latest research paper for a certain journal of Nature Publishing Group.
+-   The journals from NPG are run by different group of people, and the website of may not be consitent for all the journals
+
+</RouteEn>
 
 ### Nature Journal - News
 
@@ -37,26 +75,6 @@ pageClass: routes
 ### Nature Journal - Research Highlight
 
 <RouteEn author="yech1990" example="/nature/nature/highlight" path="/nature/nature/highlight" />
-
-### Nature Genetics (ng) - Latest Research
-
-<RouteEn author="yech1990" example="/nature/ng/research" path="/nature/ng/research" />
-
-### Nature Methods (nmeth) - Latest Research
-
-<RouteEn author="yech1990" example="/nature/nmeth/research" path="/nature/nmeth/research" />
-
-### Nature Biotechnology (nbt) - Latest Research
-
-<RouteEn author="yech1990" example="/nature/nbt/research" path="/nature/nbt/research" />
-
-### Nature Neuroscience (neuro) - Latest Research
-
-<RouteEn author="yech1990" example="/nature/neuro/research" path="/nature/neuro/research" />
-
-### Nature Machine Intelligence (natmachintell) - Latest Research
-
-<RouteEn author="LogicJake" example="/nature/natmachintell/research" path="/nature/natmachintell/research" />
 
 ### Proceedings of The National Academy of Sciences (PNAS) - Latest Articles - ALL
 
@@ -68,25 +86,35 @@ pageClass: routes
 
 ### Science Journal - Current Issue
 
-<RouteEn author="yech1990" example="/sciencemag/science/current" path="/sciencemag/science/current" />
+<RouteEn author="yech1990" example="/sciencemag/current/science" path="/nature/research/:journal" :paramsDesc="['short name for a journal']" />
+
+| `:journal` |    Full Name of the Journal    | Route                                                            |
+| :--------: | :----------------------------: | ---------------------------------------------------------------- |
+|  science   |            Science             | [/sciencemag/current/science](/sciencemag/current/science)       |
+|  advances  |        Science Advances        | [/sciencemag/current/advances](/sciencemag/current/advances)     |
+| immunology |       Science Immunology       | [/sciencemag/current/immunology](/sciencemag/current/immunology) |
+|  robotics  |        Science Robotics        | [/sciencemag/current/robotics](/sciencemag/current/robotics)     |
+|    stke    |       Science Signaling        | [/sciencemag/current/stke](/sciencemag/current/stke)             |
+|    stm     | Science Translational Medicine | [/sciencemag/current/stm](/sciencemag/current/stm)               |
+
+-   Using router (`/sciencemag/current/` + “short name for a journal”) to query current issue of a journal form AAAS.
+    leave the parameter blank（`/sciencemag/current`）to get update from Science.
+
+</RouteEn>
 
 ### Science Journal - First Release
 
-<RouteEn author="yech1990" example="/sciencemag/science/early" path="/sciencemag/science/early" />
+<RouteEn author="yech1990" example="/sciencemag/early/science" path="/sciencemag/early/science" />
 
-### Science Advances - Current Issue
+_only support Science Journal_
 
-<RouteEn author="yech1990" example="/sciencemag/advances/current" path="/sciencemag/advances/current" />
+</RouteEn>
 
 ## Search Engine
 
 ### PubMed Trending
 
 <RouteEn author="yech1990" example="/pubmed/trending" path="/pubmed/trending" />
-
-### X-MOL Platform - News
-
-<RouteEn author="cssxsh" example="/x-mol/news/3" path="/x-mol/news/:tag?" :paramsDesc="['数字编号，可从新闻列表URL得到。为空时从新闻主页获取新闻。']" />
 
 ### X-MOL Platform - Journal
