+++ docs/journal.md
@@ -10,10 +10,6 @@ pageClass: routes
 
 <Route author="yech1990" example="/pubmed/trending" path="/pubmed/trending" />
 
-### X-MOL 平台-新闻
-
-<Route author="cssxsh" example="/x-mol/news/3" path="/x-mol/news/:tag?" :paramsDesc="['数字编号，可从新闻列表URL得到。为空时从新闻主页获取新闻。']" />
-
 ### X-MOL 平台-期刊
 
 <Route author="cssxsh" example="/x-mol/paper/0/9" path="/x-mol/paper/:type/:magazine" :paramsDesc="['类别','机构，两个参数都可从期刊URL获取。']" />
@@ -41,13 +37,12 @@ pageClass: routes
 
 <Route author="yech1990" example="/cell/cell/current" path="/cell/cell/:category" />
 
-> 本期刊物 (默认选项)
+| `:category` |      类型说明       | 路由                                     |
+| :---------: | :-----------------: | ---------------------------------------- |
+|   current   | 本期刊物 (默认选项) | [/cell/cell/current](/cell/cell/current) |
+|   inpress   |      在线发表       | [/cell/cell/inpress](/cell/cell/inpress) |
 
-`/cell/cell/current`
-
-> 在线发表
-
-`/cell/cell/inpress`
+</Route>
 
 ### eLife-最新成果-综合
 
@@ -57,9 +52,49 @@ pageClass: routes
 
 <Route author="emdoe" example="/elife/cell-biology" path="/elife/:subject" :paramsDesc="['方向名称', '请在主页获取']" />
 
-### Nature 主刊-最新成果
+### Nature 系列-最新成果
 
-<Route author="yech1990" example="/nature/nature/research" path="/nature/nature/research" />
+<Route author="yech1990" example="/nature/research/ng" path="/nature/research/:journal" :paramsDesc="['期刊名简写']" />
+
+|  `:journal`   |           期刊名            | 路由                                                             |
+| :-----------: | :-------------------------: | ---------------------------------------------------------------- |
+|    nature     |           Nature            | [/nature/research/nature](/nature/research/nature)               |
+|      nbt      |    Nature Biotechnology     | [/nature/research/nbt](/nature/research/nbt)                     |
+|     neuro     |     Nature Neuroscience     | [/nature/research/neuro](/nature/research/neuro)                 |
+|      ng       |       Nature Genetics       | [/nature/research/ng](/nature/research/ng)                       |
+|      ni       |      Nature Immunology      | [/nature/research/ni](/nature/research/ni)                       |
+|     nmeth     |        Nature Method        | [/nature/research/nmeth](/nature/research/nmeth)                 |
+|     nchem     |      Nature Chemistry       | [/nature/research/nchem](/nature/research/nchem)                 |
+|     nmat      |      Nature Materials       | [/nature/research/nmat](/nature/research/nmat)                   |
+| natmachintell | Nature Machine Intelligence | [/nature/research/natmachintell](/nature/research/natmachintell) |
+
+-   通过 `/nature/research/` + “杂志简写”来获取对应杂志的最新文章（Latest Research）。
+    若参数置空（`/nature/research`），则默认获取主刊（Nature）的最新文章。
+-   由于 Nature 系列的刊物是分别由不同的编辑来独立运营，所以页面格式上有些差异。目前**仅**对以下杂志进行了测试。
+-   由于权限的限制，目前仅获取论文的摘要进行展示。
+
+</Route>
+
+### Nature-新闻及评论
+
+<Route author="yech1990" example="/nature/news-and-comment/ng" path="/nature/news-and-comment/:journal" :paramsDesc="['期刊名简写']" />
+
+|  `:journal`   |           期刊名            | 路由                                                                             |
+| :-----------: | :-------------------------: | -------------------------------------------------------------------------------- |
+|      nbt      |    Nature Biotechnology     | [/nature/news-and-comment/nbt](/nature/news-and-comment/nbt)                     |
+|     neuro     |     Nature Neuroscience     | [/nature/news-and-comment/neuro](/nature/news-and-comment/neuro)                 |
+|      ng       |       Nature Genetics       | [/nature/news-and-comment/ng](/nature/news-and-comment/ng)                       |
+|      ni       |      Nature Immunology      | [/nature/news-and-comment/ni](/nature/news-and-comment/ni)                       |
+|     nmeth     |        Nature Method        | [/nature/news-and-comment/nmeth](/nature/news-and-comment/nmeth)                 |
+|     nchem     |      Nature Chemistry       | [/nature/news-and-comment/nchem](/nature/news-and-comment/nchem)                 |
+|     nmat      |      Nature Materials       | [/nature/news-and-comment/nmat](/nature/news-and-comment/nmat)                   |
+| natmachintell | Nature Machine Intelligence | [/nature/news-and-comment/natmachintell](/nature/news-and-comment/natmachintell) |
+
+-   通过 `/nature/research/` + “杂志简写”来获取对应杂志的最新文章（Latest Research）。
+    主刊由于格式不同，该 router 并未支持，采用 `/nature/nature/news` 来获取新闻。
+-   由于 Nature 系列的刊物是分别由不同的编辑来独立运营，所以页面格式上有些差异。目前**仅**对以下杂志进行了测试。
+
+</Route>
 
 ### Nature 主刊-新闻动态
 
@@ -69,26 +104,6 @@ pageClass: routes
 
 <Route author="yech1990" example="/nature/nature/highlight" path="/nature/nature/highlight" />
 
-### Nature Genetics (ng)-最新成果
-
-<Route author="yech1990" example="/nature/ng/research" path="/nature/ng/research" />
-
-### Nature Methods (nmeth)-最新成果
-
-<Route author="yech1990" example="/nature/nmeth/research" path="/nature/nmeth/research" />
-
-### Nature Biotechnology (nbt)-最新成果
-
-<Route author="yech1990" example="/nature/nbt/research" path="/nature/nbt/research" />
-
-### Nature Neuroscience (neuro)-最新成果
-
-<Route author="yech1990" example="/nature/neuro/research" path="/nature/neuro/research" />
-
-### Nature Machine Intelligence (natmachintell)-最新成果
-
-<Route author="LogicJake" example="/nature/natmachintell/research" path="/nature/natmachintell/research" />
-
 ### PNAS-最新文章(全部)
 
 <Route author="emdoe" example="/pnas/latest" path="/pnas/latest" />
@@ -97,14 +112,28 @@ pageClass: routes
 
 <Route author="emdoe" example="/pnas/Applied Mathematics" path="/pnas/:topic" :paramsDesc="['领域名称','可从 pnas.org 获得']" />
 
-### Science 主刊-本期刊物
+### Science 系列-本期刊物
 
-<Route author="yech1990" example="/sciencemag/science/current" path="/sciencemag/science/current" />
+<Route author="yech1990" example="/sciencemag/current/science" path="/nature/research/:journal" :paramsDesc="['期刊名简写']" />
+
+| `:journal` |             期刊名             | 路由                                                             |
+| :--------: | :----------------------------: | ---------------------------------------------------------------- |
+|  science   |            Science             | [/sciencemag/current/science](/sciencemag/current/science)       |
+|  advances  |        Science Advances        | [/sciencemag/current/advances](/sciencemag/current/advances)     |
+| immunology |       Science Immunology       | [/sciencemag/current/immunology](/sciencemag/current/immunology) |
+|  robotics  |        Science Robotics        | [/sciencemag/current/robotics](/sciencemag/current/robotics)     |
+|    stke    |       Science Signaling        | [/sciencemag/current/stke](/sciencemag/current/stke)             |
+|    stm     | Science Translational Medicine | [/sciencemag/current/stm](/sciencemag/current/stm)               |
+
+-   通过 `/sciencemag/current/` + “杂志简写”来获取对应杂志最新一期的文章（Current Issue）。
+    若参数置空（`/sciencemag/current`），则默认获取主刊（Science）的最新文章。
+
+</Route>
 
 ### Science 主刊-在线发表
 
-<Route author="yech1990" example="/sciencemag/science/early" path="/sciencemag/science/early" />
+<Route author="yech1990" example="/sciencemag/early/science" path="/sciencemag/early/science" />
 
-### Science Advances-本期刊物
+_仅支持 Science 主刊_
 
-<Route author="yech1990" example="/sciencemag/advances/current" path="/sciencemag/advances/current" />
+</Route>
