+++ lib/router.js
@@ -2007,22 +2007,15 @@ router.get('/pnas/:tid', require('./routes/pnas/topic'));
 // cell [Sci Journal]
 router.get('/cell/cell/:category', require('./routes/cell/cell/index'));
 
-// nature [Sci Journal]
-router.get('/nature/nature/research', require('./routes/nature/nature/research'));
-router.get('/nature/nature/news', require('./routes/nature/nature/news'));
-router.get('/nature/nature/highlight', require('./routes/nature/nature/highlight'));
-// nature 子刊 [Sci Journal]
-router.get('/nature/nmeth/research', require('./routes/nature/nmeth/research'));
-router.get('/nature/ng/research', require('./routes/nature/ng/research'));
-router.get('/nature/nbt/research', require('./routes/nature/nbt/research'));
-router.get('/nature/neuro/research', require('./routes/nature/neuro/research'));
-router.get('/nature/natmachintell/research', require('./routes/nature/natmachintell/research'));
+// nature + nature 子刊 [Sci Journal]
+router.get('/nature/research/:journal?', require('./routes/nature/research'));
+router.get('/nature/news-and-comment/:journal?', require('./routes/nature/news-and-comment'));
+router.get('/nature/news', require('./routes/nature/news'));
+router.get('/nature/highlight', require('./routes/nature/highlight'));
 
 // science [Sci Journal]
-router.get('/sciencemag/science/current', require('./routes/sciencemag/science/current'));
-router.get('/sciencemag/science/early', require('./routes/sciencemag/science/early'));
-// science 子刊 [Sci Journal]
-router.get('/sciencemag/advances/current', require('./routes/sciencemag/advances/current'));
+router.get('/sciencemag/current/:journal?', require('./routes/sciencemag/current'));
+router.get('/sciencemag/early/science', require('./routes/sciencemag/early'));
 
 // dlsite
 router.get('/dlsite/new/:type', require('./routes/dlsite/new'));
